+++ src/settings/JobProfiles/resultsFormatter.js
@@ -1,9 +1,5 @@
 import React from 'react';
 import HighLight from 'react-highlighter';
-import {
-  get,
-  isEmpty,
-} from 'lodash';
 
 import { AppIcon } from '@folio/stripes/core';
 import {
@@ -12,12 +8,12 @@ import {
 } from '@folio/stripes/components';
 
 import {
+  TagsFormatter,
   DateFormatter,
   UserNameFormatter,
 } from '../../components';
 
 import sharedCss from '../../shared.css';
-import css from './JobProfiles.css';
 
 export const resultsFormatter = (searchTerm, selectRecord, selectedRecords) => ({
   selected: record => (
@@ -50,29 +46,7 @@ export const resultsFormatter = (searchTerm, selectRecord, selectedRecords) => (
       </HighLight>
     </AppIcon>
   ),
-  tags: record => {
-    const tags = get(record, 'tags.tagList', []);
-
-    if (isEmpty(tags)) {
-      return '-';
-    }
-
-    return (
-      <span className={css.tags}>
-        <Icon
-          size="small"
-          icon="tag"
-          iconClassName={css.tagsIcon}
-        />
-        <HighLight
-          search={searchTerm}
-          className={sharedCss.container}
-        >
-          {tags.join(', ')}
-        </HighLight>
-      </span>
-    );
-  },
+  tags: record => <TagsFormatter record={record} searchTerm={searchTerm} />,
   updated: record => {
     const { metadata: { updatedDate } } = record;