+++ src/utils/formValidators.js
@@ -2,6 +2,12 @@ import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { isEmpty } from 'lodash';
 
+/**
+ * Validates field inputs
+ *
+ * @param {string|*} value
+ * @return {null|*} Validation message
+ */
 export const validateRequiredField = value => {
   const isValid = !isEmpty(value);
 
@@ -12,6 +18,13 @@ export const validateRequiredField = value => {
   return <FormattedMessage id="ui-data-import.validation.enterValue" />;
 };
 
+/**
+ * Validates Data Types from form inputs
+ *
+ * @param {string|*} value
+ * @param {Object} formValues
+ * @return {null|*}
+ */
 export const validateDataTypes = (value, formValues) => {
   const { importBlocked } = formValues;
 
@@ -22,6 +35,12 @@ export const validateDataTypes = (value, formValues) => {
   return validateRequiredField(value);
 };
 
+/**
+ * Validates file extensions from form input
+ *
+ * @param {string|*} value
+ * @return {null|*}
+ */
 export const validateFileExtension = value => {
   const pattern = /^\.(\w+)$/;