+++ src/utils/sort.js
@@ -1,3 +1,11 @@
+/**
+ * Sorts 2 strings.
+ * Works as a sorter callback in sort() function.
+ *
+ * @param {string} a
+ * @param {string} b
+ * @return {number}
+ */
 export function sortStrings(a, b) {
   const valA = a.toUpperCase();
   const valB = b.toUpperCase();
@@ -13,6 +21,14 @@ export function sortStrings(a, b) {
   return 0;
 }
 
+/**
+ * Sorts 2 dates.
+ * Works as a sorter callback in sort() function.
+ *
+ * @param {string} a
+ * @param {string} b
+ * @return {number}
+ */
 export function sortDates(a, b) {
   const valA = new Date(a);
   const valB = new Date(b);
@@ -20,6 +36,14 @@ export function sortDates(a, b) {
   return valA - valB;
 }
 
+/**
+ * Sorts 2 numbers.
+ * Works as a sorter callback in sort() function.
+ *
+ * @param {number} a
+ * @param {number} b
+ * @return {number}
+ */
 export function sortNums(a, b) {
   return a - b;
 }